Restructured `asapdiscovery-simulation` to group all alchemy components together
Also, consolidated CLI components for `asap-alchemy` CLI into a single
file; separation didn't appear to give much advantage.
