--- asapdiscovery-simulation/asapdiscovery/simulation/cli/create.py
@@ -1,19 +1,0 @@
-import click
-
-
-@click.command()
-@click.argument(
-    "filename",
-    type=click.Path(exists=False, file_okay=True, dir_okay=False, writable=True),
-)
-def create(filename: str):
-    """
-    Create a new free energy perturbation factory with default settings and save it to JSON file.
-
-    Args:
-        filename: The name of the JSON file containing the factory schema.
-    """
-    from asapdiscovery.simulation.schema.fec import FreeEnergyCalculationFactory
-
-    factory = FreeEnergyCalculationFactory()
-    factory.to_file(filename=filename)
