+++ asapdiscovery-simulation/asapdiscovery/simulation/tests/test_utils.py
@@ -1,9 +1,9 @@
 from alchemiscale import Scope, ScopedKey
-from asapdiscovery.simulation.schema.fec import (
+from asapdiscovery.simulation.alchemy.schema.fec import (
     AlchemiscaleResults,
     FreeEnergyCalculationNetwork,
 )
-from asapdiscovery.simulation.utils import AlchemiscaleHelper
+from asapdiscovery.simulation.alchemy.utils import AlchemiscaleHelper
 from gufe.protocols import ProtocolUnitResult
 from openfe.protocols.openmm_rfe import RelativeHybridTopologyProtocolResult
 from openff.units import unit as OFFUnit
